Prevent crash caused by deleting multiple dup objects, other deleting issues
When deleting a row from a to-many subform, -removeObject was being used to remove that row from the collection. If the backing collection was a set, that was fine. However, if that backing collection was an array, then removeObject will remove all instances of that object from the array. Which could easily be referencing that object multiple times at multiple indexes. That scenario immediately crashes because the backing array would have fewer items after removeObject than the corresponding section in the tableview.

This is easy to reproduce in the sample app but just clicking add twice on the phone number section to add two blank entries and then deleting one of them. Since they are both blank, they will be -equal and the app will crash.

This also brings up the issue of what would happen if the backing store was an NSSet or NSDictionary, however. Neither guarantee the order of items returned by allObjects. Although deleting an item is unlikely to change the order of returned items, it is theoretically possible. If this did happen it would have bad consequences for the to-many list. So this commit also removes the ability to delete items from the to-many items unless the backing store is an ordered collection (NSOrderedSet or NSArray).

Conflicts:
	FXForms/FXForms.m