Use FXFormFieldInline field property to determine editability
For a to-many field, FXFormFieldInline's traditional meaning doesn't make sense: we are already, pretty much by definition, displaying the fields of the collection inline. This commit repurposes that FXFormFieldInline property for to-many fields, using it to indicate that the list of detail fields should be inline editable: allowing adds/deletes to the collection and inline editing for NSStrings.

I have very mixed feelings about this. I like neither overriding the existing meaning of FXFormFieldInline, nor do I like the fact that it doesn't provide fine grained control. But I think that displaying a read-only array of detail lines via an FXForm is a critical use case. And "allow editing inline" doesn't seem to far a stretch from "display inline".

I am happy to rework this commit for any reasonable alternative. The other alternatives I considered are:

* Add additional property types to allow fine-grained control regarding whether adds are allowed, deletes are allowed, and editing is allowed. This probably the best long-term solution, but I don't want to introduce any more changes to the properties in order to support one-to-many. I will leave any additional property decisions to @nicklockwood.

* Allow the collection type to determine whether editing inline is allowed. Only allow adds/deletes from NSMutableArray/NSMutableOrderedSet and disallow them from NSArray/NSOrderedSet. Its cognitively easy to understand/explain, but I feel that this approach violates the core design principle of FXForm: allowing you to use the objects you already have and just add a couple of methods (like -fields) to tailor what you want. This approach would make any NSArray property difficult to edit.

* Add an optional protocol method(s) to allow the user's FXForm to decide at run-time whether adds/edits/deletes are allowed. This is relatively easy to implement within FXForm, fairly easy to implement for users, and offers the most flexibility (users could allow deleting some rows, but not others). But it introduces even more complexity than adding properties, so I'm not willing to consider it seriously.